# Conflicts:
#	app/src/main/java/com/elte/recipebook/ui/screens/AddRecipeScreen.kt
#	app/src/main/java/com/elte/recipebook/viewModel/RecipeViewModel.kt

+++ app/src/main/java/com/elte/recipebook/data/entities/Nutrition.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class Nutrition (
     @PrimaryKey
     val iD: Int,
-    var energy: Double,
+    var energy: Int,
     var fat: Double,
     var protein: Double,
     var carbohydrate: Double,
